# generated/swagger-ui/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "hell manager api",
    "version" : "v1",
    "title" : "hell manager api",
    "contact" : {
      "name" : "Ots Wang",
      "url" : "http://www.otswang.win",
      "email" : "otswang@cqu.edu.cn"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host" : "localhost:8090",
  "basePath" : "/api",
...
  "paths" : {
...
    "/v0.1/security/permissions/add" : {
      "post" : {
        "tags" : [ "权限相关设置" ],
        "summary" : "新增权限",
        "description" : "",
        "operationId" : "addPermission",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "权限Vo",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/PermissionVo"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Result"
            }
          }
        }
      }
    },
    "/v0.1/security/permissions/delete" : {
      "delete" : {
        "tags" : [ "权限相关设置" ],
        "summary" : "删除权限列表, 删除顶级权限会递增删除子权限",
        "description" : "",
        "operationId" : "deletePermission",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "权限ID列表",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "integer",
              "format" : "int64"
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Result"
            }
          }
        }
      }
    },
...
    "/v0.1/security/permissions/parents/{parent_id}" : {
      "get" : {
        "tags" : [ "权限相关设置" ],
        "summary" : "获取权限列表",
        "description" : "",
        "operationId" : "getAllParentPermissions",
        "parameters" : [ {
          "name" : "parent_id",
          "in" : "path",
          "description" : "父类id",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Result"
            }
          }
        }
      }
    },
...
    "/v0.1/security/roles/add" : {
      "post" : {
        "tags" : [ "权限相关设置" ],
        "summary" : "新增角色",
        "description" : "",
        "operationId" : "addRole",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "角色Vo",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/RoleVo"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Result"
            }
          }
        }
      }
    },
    "/v0.1/security/roles/delete" : {
      "delete" : {
        "tags" : [ "权限相关设置" ],
        "summary" : "删除角色列表",
        "description" : "",
        "operationId" : "deleteRole",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "角色ID列表",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "integer",
              "format" : "int64"
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Result"
            }
          }
        }
      }
    },
    "/v0.1/security/roles/permissions/bind" : {
      "post" : {
        "tags" : [ "权限相关设置" ],
        "summary" : "更新角色权限绑定关系",
        "description" : "",
        "operationId" : "bindRolePermission",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "角色权限绑定关系",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/RelationUserRoleVo"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Result"
            }
          }
        }
      }
    },
    "/v0.1/security/users/roles/bind" : {
      "post" : {
        "tags" : [ "权限相关设置" ],
        "summary" : "更新用户角色绑定关系",
        "description" : "",
        "operationId" : "bindUserRole",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "用户角色绑定关系",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/RelationRolePermissionVo"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Result"
            }
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "PermissionVo" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "id"
        },
        "name" : {
          "type" : "string",
          "description" : "权限名称"
        },
        "parentId" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "父权限id"
        },
        "parentName" : {
          "type" : "string",
          "description" : "父权限名称"
        }
      }
    },
...
    "RelationRolePermissionVo" : {
      "type" : "object",
      "required" : [ "roleId" ],
      "properties" : {
        "roleId" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "用户id"
        },
        "permissionIds" : {
          "type" : "array",
          "description" : "角色列表，可为空",
          "uniqueItems" : true,
          "items" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      }
    },
    "RelationUserRoleVo" : {
      "type" : "object",
      "required" : [ "userId" ],
      "properties" : {
        "userId" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "用户id"
        },
        "roleIds" : {
          "type" : "array",
          "description" : "角色列表，可为空",
          "uniqueItems" : true,
          "items" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      }
    },
    "Result" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "integer",
          "format" : "int32"
        },
        "message" : {
          "type" : "string"
        },
        "data" : {
          "type" : "object"
        }
      }
    },
    "RoleVo" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "id"
        },
        "name" : {
          "type" : "string",
          "description" : "角色名称"
        },
        "code" : {
          "type" : "string",
          "description" : "角色标识"
        }
      }
    },
...
  }
}